+++ src/test/java/attendance/domain/AttendanceTest.java
@@ -1,7 +1,5 @@
-package attendance;
+package attendance.domain;
 
-import attendance.domain.Attendance;
-import attendance.domain.AttendanceStatus;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
