+++ src/test/java/attendance/domain/CrewsTest.java
@@ -1,8 +1,5 @@
-package attendance;
+package attendance.domain;
 
-import attendance.domain.Attendance;
-import attendance.domain.Crew;
-import attendance.domain.Crews;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
